move KVStore into its own package
